Code structure (#16)
* Code structure overhaul:

media.py has become too unwieldy. Individual modules will be created for track, media, video, playlist

* Split media.py into {track,video,album,playlist}.py
